small changes
Remove (float)cast only in collision.c and added comments

+++ Cproject/Src/collsion.c
@@ -28,7 +28,6 @@ void tableinit(struct wall* table, float posX, float posY)
 
 void chairinit(struct wall* chair, float posX, float posY)
 {
-
 	(*chair).x = posX;
 	(*chair).y = posY;
 	(*chair).width = 50;
@@ -38,10 +37,10 @@ void chairinit(struct wall* chair, float posX, float posY)
 }
 
 void initallwall(int numOfchair, int numOftable) {
-	float chairx[5] = { (float)((windowWidth / 2) + 200), (float)((windowWidth / 2) + 250),(float)(windowWidth / 3), 300,(float)(windowWidth - 200) };
-	float chairy[5] = { (float)((windowHeight / 2) - 50),(float)((windowHeight / 2) - 50),(float)((windowHeight - 200)), (float)((windowHeight / 2) + 50), (float)(windowHeight / 5) };
-	float tablex[3] = { (float)(windowWidth / 4),(float)((windowWidth / 2) + 200), (float)(windowWidth - 350) };
-	float tabley[3] = { (float)(windowHeight / 4), (float)(windowHeight / 2), (float)((windowHeight / 5 + 150)) };
+	float chairx[5] = { ((windowWidth / 2) + 200), ((windowWidth / 2) + 250),(windowWidth / 3), 300,(windowWidth - 200) };
+	float chairy[5] = { ((windowHeight / 2) - 50),((windowHeight / 2) - 50),((windowHeight - 200)), ((windowHeight / 2) + 50), (windowHeight / 5) };
+	float tablex[3] = { (windowWidth / 4),((windowWidth / 2) + 200), (windowWidth - 350) };
+	float tabley[3] = { (windowHeight / 4), (windowHeight / 2), ((windowHeight / 5 + 150)) };
 	for (int i = 0; i < numOfchair; i++)
 	{
 		chairinit(&chair[i], chairx[i], chairy[i]);
